C#: Assume 64-bit types when type has no meta

+++ modules/mono/editor/GodotTools/GodotTools/Export/ExportPlugin.cs
@@ -67,7 +67,7 @@ namespace GodotTools.Export
             }
         }
 
-        public override void _ExportBegin(string[] features, bool isDebug, string path, int flags)
+        public override void _ExportBegin(string[] features, bool isDebug, string path, long flags)
         {
             base._ExportBegin(features, isDebug, path, flags);
 
@@ -90,7 +90,7 @@ namespace GodotTools.Export
             }
         }
 
-        private void _ExportBeginImpl(string[] features, bool isDebug, string path, int flags)
+        private void _ExportBeginImpl(string[] features, bool isDebug, string path, long flags)
         {
             _ = flags; // Unused
